Tried to de-jargon-ify the M&S page
This page needs more work on its copy in general; we're too vague and
keep sprinkling the word "innovative" and "solution" throughout without
actually saying what we do. I think we'll need to bring in some others
to help clean this up, but I've done what I can for the moment.