(#35) - Remove excess space on blog posts
This attempts to cut down on the hero space that each page on the site
includes, since we won't create a unique hero image for each post.
However, I had to leave some space in order for the header animation to
remain properly functional. I'd suggest a better solution would be to
continue with a minimal-height hero space, but find a generic and
subtle image to reuse across posts so it doesn't look quite so bland.